Remove InterprocessTests project
I'd initially created this thinking we'd have a bunch of different test suites for different scenarios, but it hasn't really panned out that way.  In the meantime, having some tests for different projects not part of that project's test suite makes certain things more difficult, like dealing with changing dependencies, understanding code coverage, etc.  I'm simply redistributing these tests to their propery homes and deleting the consolidated project.
