--- src/Scenarios/InterProcessCommunication.sln
@@ -1,53 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.22823.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InterProcessCommunication.Tests", "tests\InterProcessCommunication\InterProcessCommunication.Tests.csproj", "{C8B651EA-21B8-45B3-9617-3BA952D0F12B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RemoteExecutorConsoleApp", "..\Common\tests\System\Diagnostics\RemoteExecutorConsoleApp\RemoteExecutorConsoleApp.csproj", "{F5E941C8-AF2F-47AB-A066-FF25470CE382}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{EFFF9E5B-58A0-4D0E-891D-40EF450FE7DA}"
-	ProjectSection(SolutionItems) = preProject
-		..\.nuget\packages.Windows_NT.config = ..\.nuget\packages.Windows_NT.config
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Linux_Debug|Any CPU = Linux_Debug|Any CPU
-		Linux_Release|Any CPU = Linux_Release|Any CPU
-		OSX_Debug|Any CPU = OSX_Debug|Any CPU
-		OSX_Release|Any CPU = OSX_Release|Any CPU
-		Windows_Debug|Any CPU = Windows_Debug|Any CPU
-		Windows_Release|Any CPU = Windows_Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C8B651EA-21B8-45B3-9617-3BA952D0F12B}.Linux_Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C8B651EA-21B8-45B3-9617-3BA952D0F12B}.Linux_Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C8B651EA-21B8-45B3-9617-3BA952D0F12B}.Linux_Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C8B651EA-21B8-45B3-9617-3BA952D0F12B}.Linux_Release|Any CPU.Build.0 = Release|Any CPU
-		{C8B651EA-21B8-45B3-9617-3BA952D0F12B}.OSX_Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C8B651EA-21B8-45B3-9617-3BA952D0F12B}.OSX_Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C8B651EA-21B8-45B3-9617-3BA952D0F12B}.OSX_Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C8B651EA-21B8-45B3-9617-3BA952D0F12B}.OSX_Release|Any CPU.Build.0 = Release|Any CPU
-		{C8B651EA-21B8-45B3-9617-3BA952D0F12B}.Windows_Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C8B651EA-21B8-45B3-9617-3BA952D0F12B}.Windows_Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C8B651EA-21B8-45B3-9617-3BA952D0F12B}.Windows_Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C8B651EA-21B8-45B3-9617-3BA952D0F12B}.Windows_Release|Any CPU.Build.0 = Release|Any CPU
-		{F5E941C8-AF2F-47AB-A066-FF25470CE382}.Linux_Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F5E941C8-AF2F-47AB-A066-FF25470CE382}.Linux_Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F5E941C8-AF2F-47AB-A066-FF25470CE382}.Linux_Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F5E941C8-AF2F-47AB-A066-FF25470CE382}.Linux_Release|Any CPU.Build.0 = Release|Any CPU
-		{F5E941C8-AF2F-47AB-A066-FF25470CE382}.OSX_Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F5E941C8-AF2F-47AB-A066-FF25470CE382}.OSX_Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F5E941C8-AF2F-47AB-A066-FF25470CE382}.OSX_Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F5E941C8-AF2F-47AB-A066-FF25470CE382}.OSX_Release|Any CPU.Build.0 = Release|Any CPU
-		{F5E941C8-AF2F-47AB-A066-FF25470CE382}.Windows_Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F5E941C8-AF2F-47AB-A066-FF25470CE382}.Windows_Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F5E941C8-AF2F-47AB-A066-FF25470CE382}.Windows_Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F5E941C8-AF2F-47AB-A066-FF25470CE382}.Windows_Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
