--- src/Scenarios/tests/InterProcessCommunication/InterProcessCommunication.Tests.csproj
@@ -1,59 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Windows_Debug</Configuration>
-  </PropertyGroup>
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.props))\dir.props" />
-  <PropertyGroup>
-    <ProjectGuid>{C8B651EA-21B8-45B3-9617-3BA952D0F12B}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <RootNamespace>InterProcessCommunication.Tests</RootNamespace>
-    <AssemblyName>InterProcessCommunication.Tests</AssemblyName>
-    <NuGetPackageImportStamp>b62eec4b</NuGetPackageImportStamp>
-  </PropertyGroup>
-  <!-- Default configurations to help VS understand the configurations -->
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Windows_Debug|AnyCPU'" />
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Windows_Release|AnyCPU'" />
-  <ItemGroup>
-    <Compile Include="AnonymousPipesTests.cs" />
-    <Compile Include="MemoryMappedFilesTests.cs" />
-    <Compile Include="NamedPipesTests.cs" />
-    <Compile Include="EventWaitHandleTests.cs" />
-    <Compile Include="SemaphoreTests.cs" />
-    <Compile Include="$(CommonTestPath)\System\IO\TempFile.cs">
-      <Link>Common\System\IO\TempFile.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonTestPath)\System\Diagnostics\RemoteExecutorTestBase.cs">
-      <Link>Common\System\Diagnostics\RemoteExecutorTestBase.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonTestPath)\System\IO\FileCleanupTestBase.cs">
-      <Link>Common\System\IO\FileCleanupTestBase.cs</Link>
-    </Compile>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="project.json" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\System.Diagnostics.Process\src\System.Diagnostics.Process.csproj">
-      <Project>{63634289-90D7-4947-8BF3-DBBE98D76C85}</Project>
-      <Name>System.Diagnostics.Process</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\..\System.IO.MemoryMappedFiles\src\System.IO.MemoryMappedFiles.csproj">
-      <Project>{16EE5522-F387-4C9E-9EF2-B5134B043F37}</Project>
-      <Name>System.IO.MemoryMappedFiles</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\..\System.IO.Pipes\src\System.IO.Pipes.csproj">
-      <Project>{63634289-90d7-4947-8bf3-dbbe98d76c85}</Project>
-      <Name>System.IO.Pipes</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\..\System.IO.UnmanagedMemoryStream\src\System.IO.UnmanagedMemoryStream.csproj">
-      <Project>{bcf9255a-4321-4277-ad7d-f5094092c554}</Project>
-      <Name>System.IO.UnmanagedMemoryStream</Name>
-    </ProjectReference>
-    <ProjectReference Include="$(CommonTestPath)\System\Diagnostics\RemoteExecutorConsoleApp\RemoteExecutorConsoleApp.csproj">
-      <Project>{69e46a6f-9966-45a5-8945-2559fe337827}</Project>
-      <Name>RemoteExecutorConsoleApp</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.targets))\dir.targets" />
-</Project>
