--- src/Scenarios/tests/InterProcessCommunication/project.json
@@ -1,22 +1,0 @@
-{
-  "dependencies": {
-    "Microsoft.NETCore.Platforms": "1.0.1-rc3-23910",
-    "System.Globalization": "4.0.10",
-    "System.IO": "4.0.10",
-    "System.IO.FileSystem": "4.0.0",
-    "System.IO.FileSystem.Primitives": "4.0.0",
-    "System.Reflection.Extensions": "4.0.0",
-    "System.Runtime": "4.0.20",
-    "System.Runtime.Extensions": "4.0.10",
-    "System.Runtime.Handles": "4.0.0",
-    "System.Threading": "4.0.10",
-    "System.Threading.Tasks": "4.0.10",
-    "xunit": "2.1.0",
-    "xunit.netcore.extensions": "1.0.0-prerelease-00187"
-  },
-  "frameworks": {
-    "dnxcore50": {
-      "imports": "portable-net45+win8"
-    }
-  }
-}
