+++ src/System.IO.Pipes/tests/project.json
@@ -3,6 +3,7 @@
     "Microsoft.DotNet.xunit.performance": "1.0.0-alpha-build0029",
     "Microsoft.NETCore.Platforms": "1.0.1-rc3-23910",
     "System.Diagnostics.Debug": "4.0.10",
+    "System.Diagnostics.Process": "4.1.0-rc3-23910",
     "System.IO": "4.0.10",
     "System.IO.FileSystem": "4.0.0",
     "System.Runtime": "4.0.20",
