+++ src/System.Threading/tests/project.json
@@ -4,6 +4,7 @@
     "Microsoft.NETCore.Platforms": "1.0.1-rc3-23910",
     "System.Collections": "4.0.10",
     "System.Diagnostics.Debug": "4.0.10",
+    "System.Diagnostics.Process": "4.1.0-rc3-23910",
     "System.Diagnostics.Tracing": "4.0.20",
     "System.Runtime": "4.0.20",
     "System.Threading": "4.0.10",
